FPLA-648 move draftwatermark generation ino existing DocmosisDocumentGeneratorService
