+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/DocmosisDocumentGeneratorService.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.reform.fpl.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,9 +16,13 @@ import uk.gov.hmcts.reform.fpl.enums.DocmosisTemplates;
 import uk.gov.hmcts.reform.fpl.model.common.DocmosisDocument;
 import uk.gov.hmcts.reform.fpl.model.common.DocmosisRequest;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Base64;
 import java.util.Map;
 
 @Service
+@Slf4j
 public class DocmosisDocumentGeneratorService {
     private final RestTemplate restTemplate;
     private final Logger logger = LoggerFactory.getLogger(getClass());
@@ -56,5 +61,16 @@ public class DocmosisDocumentGeneratorService {
         }
 
         return new DocmosisDocument(docmosisTemplate.getDocumentTitle(), response);
+    }
+
+    public String generateDraftWatermarkEncodedString() {
+        InputStream is = getClass().getResourceAsStream("/assets/images/draft-watermark.png");
+        byte[] fileContent = new byte[0];
+        try {
+            fileContent = is.readAllBytes();
+        } catch (IOException e) {
+            log.error("Unable to generate draft water image for template.", e);
+        }
+        return Base64.getEncoder().encodeToString(fileContent);
     }
 }
