+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/CaseDataExtractionServiceTest.java
@@ -33,7 +33,7 @@ import static uk.gov.hmcts.reform.fpl.utils.CaseDataGeneratorHelper.createStanda
 
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = {JacksonAutoConfiguration.class, JsonOrdersLookupService.class,
-    HearingVenueLookUpService.class, DocmosisDraftWatermarkGeneratorService.class})
+    HearingVenueLookUpService.class, DocmosisDocumentGeneratorService.class})
 class CaseDataExtractionServiceTest {
     @SuppressWarnings({"membername", "AbbreviationAsWordInName"})
 
@@ -49,7 +49,7 @@ class CaseDataExtractionServiceTest {
     private HearingVenueLookUpService hearingVenueLookUpService;
 
     @Autowired
-    private DocmosisDraftWatermarkGeneratorService draftWatermarkGeneratorService;
+    private DocmosisDocumentGeneratorService docmosisDocumentGeneratorService;
 
     private DateFormatterService dateFormatterService = new DateFormatterService();
     private HearingBookingService hearingBookingService = new HearingBookingService();
@@ -68,7 +68,7 @@ class CaseDataExtractionServiceTest {
         // required for DI
         this.caseDataExtractionService = new CaseDataExtractionService(dateFormatterService,
             hearingBookingService, hmctsCourtLookupConfiguration, ordersLookupService, directionHelperService,
-            hearingVenueLookUpService, commonCaseDataExtraction, draftWatermarkGeneratorService);
+            hearingVenueLookUpService, commonCaseDataExtraction, docmosisDocumentGeneratorService);
     }
 
     @Test
