+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/DocmosisDocumentGeneratorServiceTest.java
@@ -52,6 +52,12 @@ class DocmosisDocumentGeneratorServiceTest {
         assertThat(argumentCaptor.getValue().getBody().getOutputFormat()).isEqualTo("pdf");
     }
 
+    @Test
+    void shouldGenerateEncodedStringWhengenerateDraftWatermarkCalled() {
+        String encodedImgString = createServiceInstance().generateDraftWatermarkEncodedString();
+        assertThat(encodedImgString).isNotEqualTo("");
+    }
+
     private Map<String, Object> getTemplatePlaceholders() {
         return Map.of(
             "jurisdiction", "PUBLICLAW",
